Back End/requirements.txt
alembic==1.15.2
Faker==37.3.0
Flask==3.1.1
Flask_Bcrypt==1.0.1
flask_cors==6.0.0
Flask_JWT_Extended==4.7.1
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.31
Werkzeug==3.1.3
twilio==8.10.0
pytest-asyncio==0.22.0
